Top K Frequent Elements Solution
I struggled on this one for a bit. The struggle was how do I return the K most frequent elements once they're in a map. The best solution involves sorting them into a heap by frequency and then popping off the key for the k top elements. Since JS doesn't have a heap (and I didn't know the best solution) I spread the map into an array, sorted it in descending order, and took the first k elements of the array and returned the keys.

Time: O(nlogn)
Space: O(n)