RemoveDuplicates: attempt0 hashmap
time: O(n)
space: O(n)

Because of all the complexities of creating a map, checking conditionals, creating an array from keys, and then replacing elements in the original array we see a great example of time complexity being how a solution scales, not how fast it actually runs.

Good, but not optimal.