ProductOfArrayExceptSelf (algorithm planning)
This was a tough one! I came up with the brute force quickly, but knew it wasn't optimal. This required me to learn the prefix/suffix algorithm. The suffix was the one that caught me tripping a few times.

I have three solutions: brute force, prefix/suffix with three arrays, prefix/suffix with one array.

The last is optimal, as it optimizes for time and space complexity.