EncodeAndDecodeStrings
This one was a doozy! At first I thought, "this is the easiest medium problem I've done yet". Oh how wrong I was. Solution 0, while eloquent, isn't correct. It leaves out a ton of edge cases and naively assumes that we'll only be reading in ASCII characters- although it's got a good time/space complexity.

Solution 1 I had to learn. Encoding with the string length and a delimiter, then decoding by reading in the string length until I find a delimiter, dealing with two pointers, was just a bit hectic. I'll need to try this one again soon.