ValidParentheses: Map/stack
Time: O(n)
Space: O(n)

Optimal solution, since our chars we can get are limited and correspond we create a map out of them. Then leveraging a stack, just ensure whenever we encounter a closer that the corresponding opener is at the top of the stack. Probably could've made a better return statement but I'm alright with how I handled it.