ValidSudoku: Three Maps
Time: O(n^2)
Space: O(n^2)

Create three maps, double for loop your way through the board, create a squareKey (divide the board into 9 regions of squares). Check if each respective map has the current element (if so, it's not valid); if this element is good, add to each map if we don't have them in it and then add the current element to each map's set. That was a new one too, Sets in JS. First time for me. I struggled, will be coming back soon.